peps clean-up
