+++ yastn/tn/fpeps/ctm/__init__.py
@@ -2,5 +2,4 @@
 from ._ctmrg import ctmrg
 from ._ctm_iteration_routines import fPEPS_2layers, apply_TM_left, apply_TM_top, check_consistency_tensors
 from ._ctm_observables import nn_exp_dict, one_site_dict, measure_one_site_spin, EV2ptcorr
-from ._ctm_env import CtmEnv, init_rand, init_ones, sample, measure_2site, measure_1site
-from ._ctm_env import match_ancilla_projectors, sample_MC_
+from ._ctm_env import CtmEnv, init_rand, init_ones
