+++ yastn/tn/fpeps/gates/__init__.py
@@ -1,2 +1,1 @@
-""" imaginary time evolution routines. """
 from ._gates import match_ancilla_1s, match_ancilla_2s, gate_hopping, gate_Coulomb, gate_local_fermi_sea
